@@ -3,26 +3,16 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller as BaseController;
+use OpenApi\Attributes as OA;
 
-/**
- * @OA\Info(
- *     title=SWAGGER_APP_NAME,
- *     description=SWAGGER_APP_DESCRIPTION,
- *     version="1.0.0",
- *     @OA\Contact(
- *         name="Webuildapps",
- *         email="hello@webuildapps.com"
- *     ),
- *     @OA\License(
- *         name="MIT"
- *     )
- * )
- *
- * @OA\Server(
- *     description="Default",
- *     url=SWAGGER_APP_URL
- * )
- */
+#[OA\Info(
+    version: '1.1.0',
+    description: SWAGGER_APP_DESCRIPTION,
+    title: SWAGGER_APP_NAME,
+)]
+#[OA\Server(
+    url: SWAGGER_APP_URL,
+)]
 class Controller extends BaseController
 {
     //
