@@ -5,29 +5,27 @@ namespace App\Http\Controllers\Api;
 use App\Http\Resources\MethodResource;
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\ResourceCollection;
-use Illuminate\Validation\ValidationException;
 use Mollie\Api\Exceptions\ApiException;
 use Mollie\Laravel\Facades\Mollie;
+use OpenApi\Attributes as OA;
 
 class MethodsController extends Controller
 {
-    /**
-     * List (active) payment methods by amount.
-     *
-     * More information: https://docs.mollie.com/reference/v2/methods-api/list-methods
-     *
-     * @param Request $request
-     * @return ResourceCollection
-     * @throws ApiException
-     * @throws ValidationException
-     *
-     * @OA\Get(
-     *     path="/methods",
-     *     tags={"Method"},
-     *     @OA\Parameter(name="amount", in="query", required=false, @OA\Schema(type="number", format="double")),
-     *     @OA\Response(response=200, description="OK", @OA\JsonContent(type="array", @OA\Items(ref="#/components/schemas/Method")))
-     * )
-     */
+    #[OA\Get(
+        path: '/methods',
+        description: 'More information: https://docs.mollie.com/reference/v2/methods-api/list-methods',
+        summary: 'List (active) payment methods by amount.',
+        tags: ['Method'],
+        parameters: [
+            new OA\Parameter(name: 'amount', in: 'query', required: false, schema: new OA\Schema(type: 'number', format: 'double')),
+        ],
+        responses: [
+            new OA\Response(response: 200, description: 'Ok', content: new OA\JsonContent(properties: [
+                new OA\Property(property: 'data', type: 'array', items: new OA\Items(ref: MethodResource::class)),
+            ])),
+            new OA\Response(response: 422, description: 'Unprocessable Content', content: new OA\JsonContent),
+        ]
+    )]
     public function index(Request $request): ResourceCollection
     {
         $this->validate($request, [
