@@ -2,32 +2,25 @@
 
 namespace App\Http\Resources;
 
-use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\JsonResource;
+use OpenApi\Attributes as OA;
 
-/**
- * @OA\Schema(schema="Image")
- *
- * @OA\Property(property="size_1x", type="string", format="uri")
- * @OA\Property(property="size_2x", type="string", format="uri")
- *
- * @OA\Property(property="svg", type="string", format="uri")
- */
+#[OA\Schema(
+    schema: 'Image',
+    properties: [
+        new OA\Property('size_1x', type: 'string', format: 'uri'),
+        new OA\Property('size_2x', type: 'string', format: 'uri'),
+        new OA\Property('svg', type: 'string', format: 'uri'),
+    ]
+)]
 class ImageResource extends JsonResource
 {
-    /**
-     * Transform the resource into an array.
-     *
-     * @param Request $request
-     * @return array
-     */
-    public function toArray($request)
+    public function toArray($request): array
     {
         return [
-            'size_1x' => (string)$this->size1x,
-            'size_2x' => (string)$this->size2x,
-
-            'svg' => (string)$this->svg,
+            'size_1x' => (string) $this->size1x,
+            'size_2x' => (string) $this->size2x,
+            'svg' => (string) $this->svg,
         ];
     }
 }
